Made a util class to handle spreadsheets.
